+++ example/lib/main.dart
@@ -21,8 +21,8 @@ class _MyAppState extends State<MyApp> {
 
   // Platform messages are asynchronous, so we initialize in an async method.
   Future<void> initPlatformState() async {
-    GdprDialog().showDialog('pub id', 'https://privacy-policy/').then((onValue) {
-      print(onValue);
+    GdprDialog().showDialog('pub id', 'https://privacy-policy/', isForTest: true).then((onValue) {
+      print('result === $onValue');
     });
   }