+++ CHANGELOG.md
@@ -7,3 +7,10 @@ Write comments and release build
 
 ## 1.0.1
 Bug fix
+
+## 1.0.2
+Bug fix
+
+## 1.0.3
+Add bool variable "isForTest" for testing library, set true to activate setDebugGeography
+In release build set false or delete this argument!